response-system/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
elasticsearch==8.13.0
python-dotenv==1.0.1
pydantic==2.7.0
pytest==8.2.0
pytest-mock==3.14.0
httpx==0.27.0
